pyproject.toml
[tool.poetry]
name = "amountslicer-server-core"
version = "0.1.0"
description = "IfcOpenShell operation server of AmountSlicer"
authors = ["JakkeLab <itsjakk3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = "^0.111.0"
ifcopenshell = "^0.7.0.240627"
uvicorn = "^0.30.1"
numpy-stl = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
